feat(firefox.bash): add support for installing Firefox AppImage
This commit adds a new function called `install_appimage` to the `firefox.bash` script. This function downloads the Firefox AppImage from the official repository and installs it on the system. The function also creates a shortcut for the application in the "Network" category.

Note: The function currently only supports downloading the latest version of Firefox AppImage.